Some Debugging With Incremental Controller still needs to be tuned.

+++ 24Inch/include/IncrementalController.h
@@ -14,7 +14,7 @@ struct IncrementalController{
 
 	RedEncoder *encoder;
 	long lastReadTime;
-	int output;
+	double output;
 	double error;
 
 }typedef IncrementalController;
